Move DailyEntry around, then disable a lot of it
